api/pyproject.toml
[tool.poetry]
name = "favorite-tree-api"
version = "1.0.0"
description = "A simple API that returns my favorite tree"
authors = [
    "Mohammad Abudayyeh <mo@techtative.io>",
]
license = "All rights are reserved"
packages = [
    { include = "main" }
]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^2.1.3"
Flask-Migrate = "^2.5.3"
Flask-SQLAlchemy = "^2.4.4"
Flask-DebugToolbar = "^0.13.1"
SQLAlchemy  = "^1.3.22"
alembic = "^1.4.3"
structlog = "^20.1.0"
sqlalchemy-json = "^0.4.0"
marshmallow-sqlalchemy = "^0.28.0"
Faker = "^8.1.0"
flask-smorest = "^0.38.0"
Flask-Cors = "^3.0.10"
Flask-Seeder = "^1.2.0"
gunicorn = "^20.1.0"
json-logging = "^1.3.0"
python-json-logger = "^2.0.2"
nameparser = "^1.1.1"
marshmallow-oneofschema = "^3.0.1"
Flask-Caching = "^2.0.0"
PyMySQL = "^1.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
